feat: calculate score

+++ lib/domain/models/upgrade.dart
@@ -4,7 +4,7 @@ import 'package:easy_localization/easy_localization.dart';
 /// различных параметров игрока
 class AllUpgrade {
   final int maxLevel = 5; // максимальный уровень апгрейда
-  double score; // Начальное количество очков
+  int score; // Начальное количество очков
   double minShild = 20; // минимальный ап щита
   List<Upgrade> list = [];
